refactor: ~Data 클래스를 view.data 패키지로 이동
