+++ app/src/main/java/woowacourse/movie/view/data/LocalFormattedTime.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.view
+package woowacourse.movie.view.data
 
 import java.time.LocalTime
 import java.time.format.DateTimeFormatter
