+++ app/src/main/java/woowacourse/movie/view/data/MovieViewDatas.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.view
+package woowacourse.movie.view.data
 
 import java.io.Serializable
