+++ app/src/main/java/woowacourse/movie/view/mapper/DateRangeMapper.kt
@@ -1,7 +1,7 @@
 package woowacourse.movie.view.mapper
 
 import woowacourse.movie.domain.DateRange
-import woowacourse.movie.view.DateRangeViewData
+import woowacourse.movie.view.data.DateRangeViewData
 
 object DateRangeMapper : Mapper<DateRange, DateRangeViewData> {
     override fun DateRange.toView(): DateRangeViewData {
