+++ app/src/main/java/woowacourse/movie/view/mapper/ReservationDetailMapper.kt
@@ -2,7 +2,7 @@ package woowacourse.movie.view.mapper
 
 import woowacourse.movie.domain.Price
 import woowacourse.movie.domain.ReservationDetail
-import woowacourse.movie.view.ReservationDetailViewData
+import woowacourse.movie.view.data.ReservationDetailViewData
 
 object ReservationDetailMapper : Mapper<ReservationDetail, ReservationDetailViewData> {
     override fun ReservationDetail.toView(): ReservationDetailViewData {
