+++ app/src/main/java/woowacourse/movie/view/widget/TimeSpinner.kt
@@ -4,7 +4,7 @@ import android.os.Bundle
 import woowacourse.movie.domain.movieTimePolicy.MovieTime
 import woowacourse.movie.domain.movieTimePolicy.WeekdayMovieTime
 import woowacourse.movie.domain.movieTimePolicy.WeekendMovieTime
-import woowacourse.movie.view.LocalFormattedTime
+import woowacourse.movie.view.data.LocalFormattedTime
 import java.time.LocalDate
 
 class TimeSpinner(val spinner: SaveStateSpinner) {
